detail pg status mappings + helpdesk is state user
